requirement.txt
pandas==2.2.1
openpyxl==3.1.2
numpy>=1.24
python-docx>=1.0.0,<2.0.0